+++ src/com/efi/printsmith/pricing/copier/TotalAllCopiesPricingMethod.java
@@ -38,7 +38,7 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 				if (copierDefinition.getInterpolateMatrix() == true)
 					pricePerCopy = MatrixUtilities.calculateInterpolation(job.getPricingCopier().getCopierMatrix(), lookupQty);
 				else
-					pricePerCopy = matrixElement.getPrice1();
+					pricePerCopy = matrixElement.getPrice1().doubleValue();
 				pricePerCopy *= copierDefinition.getCopyMarkup2();
 				
 				pricingRecord.setUnitPrice(pricePerCopy + stockPrice);
@@ -48,7 +48,7 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingFirstSideRate.name())) {
 						pricingRecord.setTotalPrice((pricePerCopy * job.getTotalCopies()*2) + stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSecondSideRate.name())) {
-						pricePerSecondSide = matrixElement.getPrice2();
+						pricePerSecondSide = matrixElement.getPrice2().doubleValue();
 						pricingRecord.setUnitPrice(pricePerSecondSide);
 						pricingRecord.setTotalPrice((pricePerSecondSide * job.getTotalCopies())+ stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSideFactor.name())) {
@@ -63,8 +63,8 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 				if (copierDefinition.getInterpolateMatrix() == true)
 					discountPct = MatrixUtilities.calculateInterpolation(job.getPricingCopier().getCopierMatrix(), lookupQty);
 				else
-					discountPct = matrixElement.getPrice1();
-				pricePerCopy = copierDefinition.getBaseRate() * discountPct;
+					discountPct = matrixElement.getPrice1().doubleValue();
+				pricePerCopy = copierDefinition.getBaseRate().doubleValue() * discountPct;
 				pricePerCopy *= copierDefinition.getCopyMarkup2();
 				
 				pricingRecord.setUnitPrice(pricePerCopy + stockPrice);
@@ -74,7 +74,7 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingFirstSideRate.name())) {
 						pricingRecord.setTotalPrice((pricePerCopy * job.getTotalCopies()*2) + stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSecondSideRate.name())) {
-						pricePerSecondSide = matrixElement.getPrice2();
+						pricePerSecondSide = matrixElement.getPrice2().doubleValue();
 						pricingRecord.setUnitPrice(pricePerSecondSide);
 						pricingRecord.setTotalPrice((pricePerSecondSide * job.getTotalCopies())+ stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSideFactor.name())) {
@@ -97,7 +97,7 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingFirstSideRate.name())) {
 						pricingRecord.setTotalPrice((pricePerCopy * job.getTotalCopies()*2) + stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSecondSideRate.name())) {
-						pricePerSecondSide = matrixElement.getPrice2();
+						pricePerSecondSide = matrixElement.getPrice2().doubleValue();
 						pricingRecord.setUnitPrice(pricePerSecondSide);
 						pricingRecord.setTotalPrice((pricePerSecondSide * job.getTotalCopies())+ stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSideFactor.name())) {
@@ -112,8 +112,8 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 				if (copierDefinition.getInterpolateMatrix() == true)
 					markup = MatrixUtilities.calculateInterpolation(job.getPricingCopier().getCopierMatrix(), lookupQty);
 				else
-					markup = matrixElement.getPrice1();
-				pricePerCopy = copierDefinition.getBaseRate() * markup;
+					markup = matrixElement.getPrice1().doubleValue();
+				pricePerCopy = copierDefinition.getBaseRate().doubleValue() * markup;
 				
 				pricingRecord.setUnitPrice(pricePerCopy + stockPrice);
 				if (job.getDoubleSided()) {
@@ -122,7 +122,7 @@ public class TotalAllCopiesPricingMethod extends CopierPricingMethod {
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingFirstSideRate.name())) {
 						pricingRecord.setTotalPrice((pricePerCopy * job.getTotalCopies()*2) + stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSecondSideRate.name())) {
-						pricePerSecondSide = matrixElement.getPrice2();
+						pricePerSecondSide = matrixElement.getPrice2().doubleValue();
 						pricingRecord.setUnitPrice(pricePerSecondSide);
 						pricingRecord.setTotalPrice((pricePerSecondSide * job.getTotalCopies())+ stockPrice*job.getTotalCopies());					
 					} else if (copierDefinition.getPriceTwoSide().equals(Price2Side.UsingSideFactor.name())) {
